refactor: change Emoji to a reusable class that provides API for frontend code
